# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: VITRAS APS API
  version: "1.0.0"
  description: >
    API do sistema VITRAS APS para Atenção Primária à Saúde.
    Autenticação via JWT (Bearer) ou cookie httpOnly (vitras_access).
    Todos os endpoints autenticados exigem CSRF token (X-CSRF-Token) quando usando cookie.
    CPF e CNS em listagens são sempre mascarados (últimos 2 dígitos visíveis).
  contact:
    name: Equipe VITRAS
servers:
  - url: https://api.vitras.com.br
    description: Produção
  - url: http://localhost:3000
    description: Desenvolvimento local
security:
  - cookieAuth: []
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT obtido via /auth/login. Validade configurada (padrão: 15m)."
    cookieAuth:
      type: apiKey
      in: cookie
      name: vitras_access
      description: "Cookie httpOnly definido em /auth/login. Requer X-CSRF-Token para mutações."
  schemas:
...
    AcsVisit:
      type: object
      properties:
        id: { type: string, format: uuid }
        patientId: { type: string }
        householdId: { type: string }
        acsId: { type: string }
        visitDate: { type: string, format: date }
        outcome: { type: string }
        motivos: { type: array, items: { type: string } }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
paths:
...
  /acs-visits:
    get:
      summary: Listar visitas ACS
      tags: [ACS]
      parameters:
        - name: patientId
          in: query
          schema: { type: string }
        - name: acsId
          in: query
          schema: { type: string }
        - name: householdId
          in: query
          schema: { type: string }
        - name: startDate
          in: query
          schema: { type: string, format: date }
        - name: endDate
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Lista de visitas ACS
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AcsVisit"

    post:
      summary: Registrar visita ACS
      tags: [ACS]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patientId]
              properties:
                patientId: { type: string, format: uuid }
                householdId: { type: string }
                visitDate: { type: string, format: date }
                outcome: { type: string }
                motivos: { type: array, items: { type: string } }
                microarea: { type: string }
      responses:
        "201":
          description: Visita registrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcsVisit"
  /acs-visits/{id}:
    get:
      summary: Detalhe de visita ACS por ID
      tags: [ACS]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Visita ACS
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcsVisit"
        "404":
          description: Visita não encontrada

    patch:
      summary: Atualizar visita ACS
      tags: [ACS]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                outcome: { type: string }
                motivos: { type: array, items: { type: string } }
      responses:
        "200":
          description: Visita atualizada
        "403":
          description: Sem permissão
        "404":
          description: Visita não encontrada
        "409":
          description: Conflito de concorrência — registro foi alterado por outro usuário
  /acs-visits/patient/{patientId}:
    get:
      summary: Listar visitas por paciente
      tags: [ACS]
      parameters:
        - name: patientId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Visitas do paciente
  /acs-visits/household/{householdId}:
    get:
      summary: Listar visitas por domicílio
      tags: [ACS]
      parameters:
        - name: householdId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Visitas do domicílio
  /acs-visits/task/{taskId}:
    get:
      summary: Listar visitas por tarefa
      tags: [ACS]
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Visitas da tarefa
...